+++ .local/share/gnome-shell/extensions/just-perfection-desktop@just-perfection/lib/Prefs/Prefs.js
@@ -2,7 +2,7 @@
  * Prefs Library
  *
  * @author     Javad Rahmatzadeh <j.rahmatzadeh@gmail.com>
- * @copyright  2020-2022
+ * @copyright  2020-2023
  * @license    GPL-3.0-only
  */