+++ .local/share/gnome-shell/extensions/just-perfection-desktop@just-perfection/CHANGELOG.md
@@ -6,6 +6,32 @@ We go to the next version after each release on [GNOME Shell Extensions website]
 
 ## [Unreleased]
 
+## [23.0.0 Goya] - 2023-02-10
+
+### Changed
+
+- Clock menu offset maximum to 20.
+- Fastest animation speed (twice as fast before).
+
+### Fixed
+
+- App menu label visibility compatibility issue with open whether extension.
+- Hidden workspace switcher padding in overview.
+- Keep clock menu position when another widget adds to the panel.
+- Looking glass API conflict with Dash to Panel extension.
+- Looking glass error when primary monitor doesn't exist.
+- Window preview gap when window picker close button is disabled.
+- Workspace boxes top margin in app grid when the search entry is disabled.
+- Italian translation by [@svityboy](https://gitlab.gnome.org/svityboy).
+
+### Added
+
+- More dash icon size (16px, 22px, 24px, 40px and 56px).
+- Overview spacing size.
+- Screen recording indicator visibility.
+- Screen sharing indicator visibility.
+- Switcher popup delay by [@koplo199](https://gitlab.gnome.org/koplo199).
+
 ## [22.0.0 Millet] - 2022-09-10
 
 ### Fixed
@@ -417,4 +443,4 @@ We go to the next version after each release on [GNOME Shell Extensions website]
 - Hide workspace switcher.
 - Hot corner to toggle overview visibility.
 
-.
+